Set the default 'updateStrategy' to 'RollingUpdate'
This change allows us to patch the device plugin to (for example) update the
image tag to the latest version and have an automated rolling upgrade be
performed to update the plugin on all agent nodes.

Previously, the device plugin on each node would *only* be restarted with the
new image if the pod representing it was explicitly shutdown and restarted.