Introduce a ResourceManager() interface to simplify resource management
This abstraction encapsulates the functions needed to get a list of
devices for a device plugin, as well as perform health checks on them.

At present, there is only one implementation of this interface (the
GpuDeviceManager) and it is associated with the "nvidia.com/gpu"
resource type when instantiating a plugin.

Once the plugin is instantiated with a specific ResourceManager, it
becomes a ResourceManager itself and can call out to the ResourceManager
interfaces on itself, or be pased along to anything else that expects a
ResourceManager interface.

Signed-off-by: Kevin Klues <kklues@nvidia.com>