Update plugin logic so that it can be started and stopped reliably
Previously, we had to instantiate a new instance of an NvidiaDevicePlugin
anytime we wanted to "restart" it. The logic has been updated to allow a
stop and restart on an existing instance of NvidiaDevicePlugin(). This
simplifies the logic in maing.go significantly since we now only have to
instantiate each of our plugins once.

Signed-off-by: Kevin Klues <kklues@nvidia.com>