Added a helm chart for the nvidia-device-plugin daemonset
A set of flags control how the YAML will be generated when installing
via helm:

```
  legacyDaemonsetAPI
        indicates whether we want to use the legacy daemonset API version 'extensions/v1beta1' or not
        (default 'false')
  compatWithCPUManager
        indicates whether we run with escalated privileges to be compatible with the CPUManager or not
        (default 'false')
```

The `compatWithCPUManager` flag configures the daemonset to be able to
interoperate with the static `CPUManager` of the `kubelet`.  Setting this flag
requires one to deploy the daemonset with elevated privileges, so only do so if
you know you need interopraability with the `CPUManager`.

The `legacyDaemonsetAPI` flag configures the daemonset to use version
`extensions/v1beta1` of the DaemonSet API. This API version was removed
in Kubernetes `v1.16`, so is only intended to allow the latest plugin to
be run on older versions of Kubernetes.

We also allow overrides for the following common user-specific settings:
- namespace
- nodeSelector
- affinity
- tolerations

A future commit will udate the README with details of how to use this.

Signed-off-by: Kevin Klues <kklues@nvidia.com>