Add best effort remove off socket file before serving
Previously, we would run into issues where this file wasn't properly
removed if the plugin crashed or was killed out-of band via a SIGKILL.
When the prlugin restarted, it would then sit in an infinite loop
warning us that it couldnÄt start because the socket file already
existed.

This patch adds a best-effort removal of the socket file before
attempting to serve it. There is never a case where we want multiple
plugins serving on the same socket file, so this should be safe to do.