Fail the plugin if NVML cannot be loaded
Previously, the plugin would not error out if NVML could not be loaded.
Instead, it would sit on an empty select{} with an error message printed
to the logs as to why the plugin wasn't able to move forward.

This was originally intended to allow the plugin to deploy successfully
on nodes that didn't have GPUs on them (and weren't supposed to have
GPUs on them) without throwing an error. In this way, you could blindly
deploy a daemonset with the plugin on all nodes in your cluster, whether
they had GPus on them or not, without encountering an error. However,
doing so meant that there was no way to detect an actual error with the
deployment on nodes that were supposed to have GPUs.

This patch updates this logic to print the same error message (plus a
little more information), but also error out instead of simply blocking
indefinitely when loading NVML fails.

This patch introduces a user-facing change, such that if the plugin is
deployed on nodes that do not have GPUs on them, it will now start to
fail (where it didn't in the past). Users need to update their
deployments (and their cluster configuration as necessary) to ensure
that the plugin only lands on nodes that have GPUs installed on them.
The typical way to do this is through a toleration or a nodeSelector.
Setting up a taint on GPU nodes for 'nvidia.com/gpu' can help with the
former. Node feature discovery (NFD) and GPU feature discovery (GFD)
can help with the latter.

A flag has been added to preserve the previous semantics for the time
being as well. However, must be explicitly turned on and is no longer the
default.

Signed-off-by: Kevin Klues <kklues@nvidia.com>