Allow mig.strategy=single on nodes with non-MIG gpus
This change allows the mig.strategy=single label to be applied to
nodes with no MIG devices. In this case, the full GPUs are exposed in
the same way as selecting mig.strategy=none.

This also handles the case that was missed in the past where devices
with migEnabled=false were ignored when checking for the applicability
of mig.strategy=single

Signed-off-by: Evan Lezar <elezar@nvidia.com>