Add a flag to specify the root path of the NVIDIA driver installation
The --pass-device-specs flag always assumes a driver installed at `/`.
When running with the operator, however, the driver is installed at
`/run/nvidia/driver`, causing this feature to fail when enabled in the
GPU Operator.

This patch adds a new flag to specify the root of the driver
installation so that the plugin can adjust the path of the /dev nodes it
injects.

Signed-off-by: Kevin Klues <kklues@nvidia.com>