Add a new --device-id-strategy flag
The `device-id-strategy` flag allows one to choose which strategy the
plugin will use to pass the device ID of the GPUs allocated to a
container. The device ID has traditionally been passed as the UUID of
the GPU. This flag lets a user decide if they would like to use the UUID
or the index of the GPU (as seen in the output of `nvidia-smi`) as the
identifier passed to the underlying runtime.  Passing the index may be
desirable in situations where pods that have been allocated GPUs by the
plugin get restarted with different physical GPUs attached to them.

Signed-off-by: Kevin Klues <kklues@nvidia.com>