accepting a XIDs to ignore from healthcheck
It's not always clear if the XID reported by the hardware is actual
hardware or application errors. The list if XID is also not alway up to
date. Allowing the end user to provide a list of XIDs they can safely
ignore to prevent GPUs from become unhealthy when it's an application
issues.