Allow skipped Xids to be specified in DP_DISABLE_HEALTHCHECKS envvar
This change allows additional XIDs to be skipped to be specified in the
DP_DISABLE_HEALTHCHECKS envvar. If the envvar has the value "all" or
CONTAINS the string "xids", the healtchecks will be disabled entirely.
(This matches the existing behaviour).

If these conditions are not met, the envar is treated as a comma-separated
list of XIDs to ignore. These are ignored in addition to the list of
application error XIDs to ignore and allows an operator to skip additional
use-case specific Xids.

Signed-off-by: Evan Lezar <elezar@nvidia.com>