Add internal 'ResourceManager' package
This package is responsible to reading the 'Resources' section of the config
file and generating ResourceManagers for each of the resources matched in
there. For backwards compatibility, if devices exist that do not match and
selectors in the 'Resource' config, then we assign these devices to the same
resources they would have been assigned to previously (depending on the MIG
strategy and type of device).

Signed-off-by: Kevin Klues <kklues@nvidia.com>