Update config file to be used across plugin and GFD
In general, its a little weird to have the combined config live in just one of
the two repos, but since these components are so tightly coupled, it's not too
terrible. In the future we should consider merging the plugin and GFD into the
same code base, but continue to maintain separate binaries and helm charts /
daemonset yamls. It probably makes the most sense if/when we move to calendar
versioning instead of semantic versioning for these components.

Signed-off-by: Kevin Klues <kklues@nvidia.com>