Update device.IDs sent to the kubelet to be of the form <id>::<replica>
For now, all devices are explicity set to have 1 replica, but a subsequent
commit will allow them to be oversubscribed with more replicas. The code here
prepares the interaction with the kubelet (i.e. adding the replicas and
stripping from the IDs as necessary) to be verified before adding any more
sophisticated logic around generating the replicas.

Signed-off-by: Kevin Klues <kklues@nvidia.com>