Remove timeSlicing.strategy and any allocation policies built around it.
After some discussion, it doesn't make sense to have a knob to control this as
originally intended, because a strict packed vs. distributed allocation
strategy is not likely something people will want in practice. In practice, a
strategy that packs allocations within a single container, but distributes them
across containers is more desirable. We will revisit this in the future, and
likely not even include a knob for it, but instead just make it the default /
only way of doing things.

Signed-off-by: Kevin Klues <kklues@nvidia.com>