Blindly move gfd daemonset.yml to gfd.yaml in plugin helm chart
This is a bit unorthodox because (logically) the gfd daemonset template should
live inside the gfd chart. However, there is currently no good way to pass
values from a parent chart into a subchart without jumping through alot of
undesirabe hoops. This makes it hard to support the new shared config we have
between gfd and the plugin.

By moving the gfd daemonset into the plugin chart, we now have the ability to
share values from the plugin into gfd. Note, that we don't move the contents of
the values.yml from the gfd subchart into the plugin's values.yml. This is done
on purpose, so that once proper suppport is in place to share values from a
parent to a subchart, there won't be much work to move this file back where it
belongs. A subsequent commit will show how we access these subchart values from
gfd.yml, such that minimal code changes will be needed once we move gfd.yml
back into the subchart.

Below is a link to the Helm Improvement Proposal (HIP) that (once available)
will allow us to move gfd.yml back where it belongs:

https://github.com/helm/community/pull/242

Signed-off-by: Kevin Klues <kklues@nvidia.com>