Blindly move nfd as a subchart of gfd to a subchart of the plugin
As part of this, we rename the subchart "condition" to nfd.enabled instead of
nfd.deploy to be more in line with standard practices.

See the following for more details:
https://helm.sh/docs/chart_best_practices/dependencies/#conditions-and-tags

We also made 'nfd' an alias for the subchart so that users can set nfd
variables through 'nfd.<field>' instead of having to write out
'node-feature-discovery.<field>'.

Signed-off-by: Kevin Klues <kklues@nvidia.com>