Handle cases where no default config is specified to config-manager
In this case it looks for a config called 'default' to be available in the list
of files mounted as part of the config map. If that is not found, it just picks
an arbitrary config file (we choose to pick the file with the "largest" name
sorted alphabetically). Sorting this way ensures we get the same file on all
nodes, and avoids the "special" ..data files inserted as part of the configmap.

Signed-off-by: Kevin Klues <kklues@nvidia.com>