Add cdi-annotations DeviceListStrategy
This change adds a cdi-annotations DeviceListStrategy that specifies
the required devices through CDI annotations for the container. These
are interpreted by a CDI-enabled container engine (or the NVIDIA
Container Runtime) and compared to a CDI specification to make the
required modifications to the OCI runtime specification.

Note that if GDS and MOFED are enabled, the nvidia.com/gds=all
and nvidia.com/mofed=all devices are also requested.

Signed-off-by: Evan Lezar <elezar@nvidia.com>