Add static deployement file for Microshift
I have done some test with Microshift and the device plugin deployment
requires some tweaks, such as running it with the privileged security
context [1] to be able to mount the `/var/lib/kubelet/device-plugins`
host folder via a `hostPath` volume.

In turn, this is better to make it run in its own namespace, with it own
service account. So, the deployment file creates them and the required
role and role binding.

The `PASS_DEVICE_SPECS=true` environment variable is also set to be
compatible with CPU Manager.

[1] https://kubernetes.io/docs/concepts/extend-kubernetes/compute-storage-net/device-plugins/#device-plugin-deployment

Signed-off-by: Fabien Dupont <fdupont@redhat.com>