Add nvidiaDevRoot config option
This change adds a nvidaDevRoot config option to allow for driver
installations where the device nodes are not rooted at the nvidiaDriverRoot.

This option can be controlled using the HOST_DEV_ROOT environment variable
or host-dev-root command line argument.

If this value is not specified, the value of nvidiaDriverRoot will be used.

Signed-off-by: Evan Lezar <elezar@nvidia.com>