Add optional injection of the default IMEX channel
This change adds the ability to inject the default IMEX channel.
This is controlled by a config or comand line options which supports
one of three values. "disabled" (default), "enabled", "auto".

When enabled, the default IMEX channel is added to the container allocate
response by the mechanism appropriate for the selected deviceListStrategies.

If the defaultChannelStrategy is configured to "auto" the addition of the
IMEX channel to the containerAllocateResponse is only triggered if the
associated device node exists, otherwise the behaviour is equivalent to
"disabled"

Signed-off-by: Evan Lezar <elezar@nvidia.com>