Add init container to make imex nodes config available to container
This change adds an init container to GFD that checks whether the
/driver-root/etc/nvidia-imex/nodes_config.cfg file exits and copies
it to /config/etc/nvidia-imex/nodes_config.cfg (where /config is
a shared volume used by the containers in the GFD pod.

This allows the main GFD container access to this file -- used for generating
IMEX domain labels -- without requiring continuouse. access to the driver root.
Note that since the file is copied, a change in this config file requires a restart
of the GFD pod.

Signed-off-by: Evan Lezar <elezar@nvidia.com>