bugfix: Allow for optional timeslicing configuration
This fixes a vestigial bug from the introduction of MPS. Now that two timeslicing configurations are available, it's possible for MPS to be configured instead of timeslicing.

But since the TimeSlicing field was made non-optional - its existence is forced when unmarshalling, and then the parsing fails because no resources are specified under the empty TimeSlicing field.

Signed-off-by: Daniel Kleinstein <daniel.kleinstein@gmail.com>