Add support for explicitly enabling XIDs in health checks
This change adds support for users to explicitly set specific XIDs as fatal
errors that cause a device to be marked as unhealth. When used in conjunction
with the ability to ignore ALL XIDs, the set of XIDs that are fatal can be set.

To do this, an environment variable DP_ENABLE_HEALTHCHECKS that mirrors the
existing DP_DISABLE_HEALTHCHECKS envvar is added.

Explicitly enabling an XID overrides disabling them.

Signed-off-by: Robert Smith <robertdavidsmith@yahoo.com>
Co-Authored-by: Evan Lezar <elezar@nvidia.com>